+++ schemes/bgv/bgv_test.go
@@ -668,77 +668,92 @@ func testEvaluatorBvg(tc *TestContext, t *testing.T) {
 	}
 
 	// Naive implementation of the inner sum for reference
-	innersum := func(values []uint64, n, batchSize int) {
-		tmp := make([]uint64, len(values))
-		copy(tmp, values)
+	innersum := func(values []uint64, n, batchSize int, rotateAndAdd bool) {
+		aggregate := false
+		if n*batchSize == len(values) && !rotateAndAdd {
+			aggregate = true
+			n = n / 2
+		}
+		halfN := len(values) >> 1
+		tmp1 := make([]uint64, halfN)
+		tmp2 := make([]uint64, halfN)
+		copy(tmp1, values[:halfN])
+		copy(tmp2, values[halfN:])
 		for i := 1; i < n; i++ {
-			rot := utils.RotateSlice(tmp, i*batchSize)
-			for j := range values {
-				values[j] = (values[j] + rot[j]) % tc.Params.PlaintextModulus()
+			rot1 := utils.RotateSlice(tmp1, i*batchSize)
+			rot2 := utils.RotateSlice(tmp2, i*batchSize)
+			for j := range rot1 {
+				values[j] = (values[j] + rot1[j]) % tc.Params.PlaintextModulus()
+				values[j+halfN] = (values[j+halfN] + rot2[j]) % tc.Params.PlaintextModulus()
+			}
+		}
+		if aggregate {
+			for i := range tmp1 {
+				values[i] = (values[i] + values[i+halfN]) % tc.Params.PlaintextModulus()
 			}
 		}
 	}
 
-	for _, N := range []int{tc.Params.N(), tc.Params.MaxSlots()} {
-		for _, lvl := range testLevel {
-			t.Run(name("Evaluator/InnerSum/N slots", tc, lvl), func(t *testing.T) {
-				if lvl == 0 {
-					t.Skip("Skipping: Level = 0")
-				}
-				n := N >> 2
-				batchSize := 1 << 2
+	for _, i := range []int{0, 1, 2} {
+		// n*batchSize = N, N/2, N/8
+		for _, offset := range []int{0, 1, 3} {
+			for _, lvl := range testLevel {
+				t.Run(name("Evaluator/InnerSum/", tc, lvl), func(t *testing.T) {
+					if lvl == 0 {
+						t.Skip("Skipping: Level = 0")
+					}
+					n := tc.Params.MaxSlots() >> (i + offset)
+					batchSize := 1 << i
 
-				galEls := tc.Params.GaloisElementsForInnerSum(batchSize, n)
-				evl := tc.Evl.WithKey(rlwe.NewMemEvaluationKeySet(nil, tc.Kgen.GenGaloisKeysNew(galEls, tc.Sk)...))
+					galEls := tc.Params.GaloisElementsForInnerSum(batchSize, n)
+					evl := tc.Evl.WithKey(rlwe.NewMemEvaluationKeySet(nil, tc.Kgen.GenGaloisKeysNew(galEls, tc.Sk)...))
 
-				want, _, ciphertext0 := NewTestVector(tc.Params, tc.Ecd, tc.Enc, lvl, tc.Params.NewScale(3))
+					want, _, ciphertext0 := NewTestVector(tc.Params, tc.Ecd, tc.Enc, lvl, tc.Params.NewScale(3))
 
-				innersum(want, n, batchSize)
+					innersum(want, n, batchSize, false)
 
-				receiver := NewCiphertext(tc.Params, 1, lvl)
+					receiver := NewCiphertext(tc.Params, 1, lvl)
 
-				require.NoError(t, evl.InnerSum(ciphertext0, batchSize, n, receiver))
+					require.NoError(t, evl.InnerSum(ciphertext0, batchSize, n, receiver))
 
-				have := make([]uint64, len(want))
-				require.NoError(t, tc.Ecd.Decode(tc.Dec.DecryptNew(receiver), have))
+					have := make([]uint64, len(want))
+					require.NoError(t, tc.Ecd.Decode(tc.Dec.DecryptNew(receiver), have))
 
-				for i := 0; i < len(want); i += n * batchSize {
-					require.Equal(t, want[i:i+batchSize], have[i:i+batchSize])
-				}
-			})
+					for i := 0; i < len(want); i += n * batchSize {
+						require.Equal(t, want[i:i+batchSize], have[i:i+batchSize])
+					}
+				})
+			}
 		}
-	}
 
-	for _, lvl := range testLevel {
-		t.Run(name("Evaluator/InnerSum/N/2 slots", tc, lvl), func(t *testing.T) {
-			if lvl == 0 {
-				t.Skip("Skipping: Level = 0")
+		// Test RotateAndAdd with n*batchSize dividing and not dividing #slots
+		for _, n := range []int{tc.Params.MaxSlots() >> 3, 7} {
+			for _, batchSize := range []int{8, 3} {
+				for _, lvl := range testLevel {
+					t.Run(name("Evaluator/RotateAndAdd/", tc, lvl), func(t *testing.T) {
+						if lvl == 0 {
+							t.Skip("Skipping: Level = 0")
+						}
+
+						galEls := tc.Params.GaloisElementsForInnerSum(batchSize, n)
+						evl := tc.Evl.WithKey(rlwe.NewMemEvaluationKeySet(nil, tc.Kgen.GenGaloisKeysNew(galEls, tc.Sk)...))
+
+						want, _, ciphertext0 := NewTestVector(tc.Params, tc.Ecd, tc.Enc, lvl, tc.Params.NewScale(3))
+
+						innersum(want, n, batchSize, true)
+
+						receiver := NewCiphertext(tc.Params, 1, lvl)
+
+						require.NoError(t, evl.RotateAndAdd(ciphertext0, batchSize, n, receiver))
+
+						have := make([]uint64, len(want))
+						require.NoError(t, tc.Ecd.Decode(tc.Dec.DecryptNew(receiver), have))
+
+						require.Equal(t, want, have)
+					})
+				}
 			}
-			n := 7
-			batchSize := 13
-			l := n * batchSize
-			halfN := tc.Params.MaxSlots() >> 1
-
-			galEls := tc.Params.GaloisElementsForInnerSum(batchSize, n)
-			evl := tc.Evl.WithKey(rlwe.NewMemEvaluationKeySet(nil, tc.Kgen.GenGaloisKeysNew(galEls, tc.Sk)...))
-
-			want, _, ciphertext0 := NewTestVector(tc.Params, tc.Ecd, tc.Enc, lvl, tc.Params.NewScale(3))
-
-			innersum(want[:halfN], n, batchSize)
-			innersum(want[halfN:], n, batchSize)
-
-			receiver := NewCiphertext(tc.Params, 1, lvl)
-
-			require.NoError(t, evl.InnerSum(ciphertext0, batchSize, n, receiver))
-
-			have := make([]uint64, len(want))
-			require.NoError(t, tc.Ecd.Decode(tc.Dec.DecryptNew(receiver), have))
-
-			for i, j := 0, halfN; i < halfN; i, j = i+l, j+l {
-				require.Equal(t, want[i:i+batchSize], have[i:i+batchSize])
-				require.Equal(t, want[j:j+batchSize], have[j:j+batchSize])
-			}
-		})
+		}
 	}
 }